refactor: rename Select component
